Rename FolderBrowserPathNavigator to FolderBrowserPathView #5557
