+++ ui/app/src/utils/state.ts
@@ -126,16 +126,12 @@ export function getStoredPathNavigatorTree(siteIdentifier: string, user: string,
   return JSON.parse(window.localStorage.getItem(`craftercms.${user}.pathNavigatorTree.${siteIdentifier}.${id}`));
 }
 
-export function setStoredFolderBrowserPathNavigator(
-  siteIdentifier: string,
-  user: string,
-  value: { limit: number }
-): void {
-  window.localStorage.setItem(`craftercms.${user}.folderBrowserPathNavigator.${siteIdentifier}`, JSON.stringify(value));
+export function setStoredFolderBrowserPathView(siteIdentifier: string, user: string, value: { limit: number }): void {
+  window.localStorage.setItem(`craftercms.${user}.folderBrowserPathView.${siteIdentifier}`, JSON.stringify(value));
 }
 
-export function getStoredFolderBrowserPathNavigator(siteIdentifier: string, user: string) {
-  return JSON.parse(window.localStorage.getItem(`craftercms.${user}.folderBrowserPathNavigator.${siteIdentifier}`));
+export function getStoredFolderBrowserPathView(siteIdentifier: string, user: string) {
+  return JSON.parse(window.localStorage.getItem(`craftercms.${user}.folderBrowserPathView.${siteIdentifier}`));
 }
 
 export function setStoredGlobalAppOpenSidebar(user: string, value: boolean) {
